+++ CHANGELOG.md
@@ -2,7 +2,6 @@
 
 The changelog for `Superwall`. Also see the [releases](https://github.com/superwall/Superwall-Android/releases) on GitHub.
 
-
 ## 1.1.9
 
 ### Deprecations
@@ -12,7 +11,7 @@ The changelog for `Superwall`. Also see the [releases](https://github.com/superw
 ### Fixes
 
 - Fixes SW-2878 and it's related leaks. The `PaywallViewController` was not being properly detached when activity was stopped, causing memory leaks.
-
+- SW-2872: Fixes issue where `deviceAttributes` event and fetching would not await for IP geo to complete.
 
 ## 1.1.8